Move 'getNumberedSeries' to material/helper.ts
it is both used in the search result app and the material app, but it is mostly material related
