+++ src/apps/material/helper.ts
@@ -350,3 +350,18 @@ export const getBestMaterialTypeForWork = (work: Work) => {
 export const reservationModalId = (faustIds: FaustId[]) => {
   return constructModalId("reservation-modal", faustIds.sort());
 };
+
+export const getNumberedSeries = (series: Work["series"]) => {
+  // Filter out series that don't have a 'numberInSeries' property
+  const seriesWithNumber = series.filter(
+    (serie) => serie.numberInSeries?.number
+  );
+
+  // Sort the series based on their 'numberInSeries' property
+  return seriesWithNumber.sort((a, b) => {
+    if (a?.numberInSeries?.number && b?.numberInSeries?.number) {
+      return Number(a.numberInSeries.number) - Number(b.numberInSeries.number);
+    }
+    return 0;
+  });
+};
