+++ src/core/utils/helpers/general.ts
@@ -377,19 +377,4 @@ export const filterLoansNotOverdue = (loans: LoanType[], warning: number) => {
 export const constructModalId = (prefix: string, fragments: string[]) =>
   `${prefix ? `${prefix}-` : ""}${fragments.join("-")}`;
 
-export const getNumberedSeries = (series: Work["series"]) => {
-  // Filter out series that don't have a 'numberInSeries' property
-  const seriesWithNumber = series.filter(
-    (serie) => serie.numberInSeries?.number
-  );
-
-  // Sort the series based on their 'numberInSeries' property
-  return seriesWithNumber.sort((a, b) => {
-    if (a?.numberInSeries?.number && b?.numberInSeries?.number) {
-      return Number(a.numberInSeries.number) - Number(b.numberInSeries.number);
-    }
-    return 0;
-  });
-};
-
 export default {};
